pyproject.toml
[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"

[project]
name = "atlascloud_comfyui"
version = "0.0.1"
description = "A collection of custom nodes for ComfyUI"
authors = [
  {name = "AtlasCloud", email = "irene.chen@atlascloud.ai"}
]
readme = "README.md"
license = {text = "MIT license"}
requires-python = ">=3.10"
classifiers = []
dependencies = [
    "requests",
    "numpy",
    "pillow"
]



[project.optional-dependencies]
dev = [
    "bump-my-version",
    "coverage",  # testing
    "mypy",  # linting
    "pre-commit", # runs linting on commit
    "pytest",  # testing
    "ruff",  # linting
]

[project.urls]
Repository = "https://github.com/AtlascloudAI/atlascloud_comfyui"
BugTracker = "https://github.com/AtlascloudAI/atlascloud_comfyui/issues"
Documentation = "https://github.com/AtlascloudAI/atlascloud_comfyui/wiki"


[tool.comfy]
PublisherId = "YC0101"
DisplayName = "AtlasCloud_ComfyUI"
Icon = ""
Tags = []
Repository = "https://github.com/AtlascloudAI/atlascloud_comfyui"

includes = ["src/atlascloud_comfyui"]

[tool.setuptools.package-data]
"*" = ["*.*"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = [
    "tests",
]

[tool.mypy]
files = "."

# Use strict defaults
strict = true
warn_unreachable = true
warn_no_return = true

[[tool.mypy.overrides]]
# Don't require test functions to include types
module = "tests.*"
allow_untyped_defs = true
disable_error_code = "attr-defined"

[[tool.mypy.overrides]]
# Allow untyped imports for third-party libraries
module = "atlascloud_comfyui.client.*"
ignore_missing_imports = true

[tool.ruff]
# extend-exclude = ["static", "ci/templates"]
line-length = 140
src = ["src", "tests"]
target-version = "py39"

# Add rules to ban exec/eval
[tool.ruff.lint]
select = [
    "S102",  # exec-builtin
    "S307",  # eval-used
    "W293",
    "F",  # The "F" series in Ruff stands for "Pyflakes" rules, which catch various Python syntax errors and undefined names.
    # See all rules here: https://docs.astral.sh/ruff/rules/#pyflakes-f
]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"
